as type int removed

+++ riskaverse_corrected/risk_averse_rvns_db_holding_cost_sensitivity.py
@@ -117,7 +117,7 @@ def optimal_server_number2(priority, FailureRates, ServiceRates, holding_costs, 
             TotalCost = temp_TotalCost
             Server_Cost = temp_Server_Cost
             holdList, backorderList, SList = temp_holdList, temp_backorderList, temp_SList
-    return Min_server, min_nserver, TotalCost, Server_Cost, np.mean(holdList), np.mean(backorderList), np.mean(SList, axis=0).astype(int)
+    return Min_server, min_nserver, TotalCost, Server_Cost, np.mean(holdList), np.mean(backorderList), np.mean(SList, axis=0)
 
 
 def set_multipliers(len_sku):
@@ -451,8 +451,6 @@ for case in json_case[0]:
                     VNS_SimOpt["Utilization"] = np.sum(FailureRates/ServiceRates) / used_ser
                     VNS_SimOpt["Num_back_orders"] = bo_cost / float(penalty_cost)
                     VNS_SimOpt["S"] = S.tolist()
-
-
 
 
                     Results.append(VNS_SimOpt)
